zhangsx 2016/3/22 11:26

+++ app/src/main/java/com/handsome/qhb/bean/User.java
@@ -41,7 +41,7 @@ public class User implements Serializable {
         return integral;
     }
 
-    public void setIntegral(int integral) {
+    public void setIntegral(float integral) {
         this.integral = integral;
     }
